fix: use clang-11 consistently
In 42ba4b5a4 or #671 llvmPackages_11 was added to the build,
presumably to help with cross compiling. However, the inclusion of
x86intrin.h via nvme_internal.h when building spdk-sys resulted in
"invalid conversion" errors in the clang headers when running nix
build to build the mayastor image, but not in the regular nix-shell
environment.

That may be due to the use of clang-11 headers with clang-7 libraries
so use clang_11 and llvmPackages_11 throughout, update LIBCLANG_PATH,
and drop llvmPackages.

Fixes: 42ba4b5a4 ("chore: try to enable cross builds")