feat(moac): switch to grpc-js from grpc-uds
The workaround for UDS sockets in grpc lib is no longer needed so we
can get rid of grpc library that is obsolete and unmaintained.  It
has been tested with nodejs v16 + k8s 1.20.7 with current versions of
csi-attacher and csi-provisioner.

It is also a big time saver when building moac image as we don't have
to compile the C++ grpc library in nix anymore.

grpc library in mayastor grpc tests stays and in addition to that we
use grpc-js as well. We will revisit that in the future and convert
everything to grpc-js. That has lower priority as it is just a test
code.

CSI sidecars used with moac have been updated to the latest version.
We need to fetch the images from k8s.gcr.io registry rather than from
quay that does not have the latest versions.

Resolves: CAS-927