fix: more Initiator IO failures
Defer failing the IO on more paths but open the fail tap again once we
retire the child before pausing and the pausing op cannot complete if
there are outstanding NS IO's.

This means there is still a small gap for IO failures but in practice that'll
probably only cause at most 1 more failure which is below the 5 limit for the initiator.